Crucial (and untested) changes to main.ts
Command selection now uses a "switch" instead of a bunch of "else if" (was it a necessary change? maybe. did I just want to change something so I could contribute? yes.)
Also, added the ping value to the "ping" command, however this is more of an estimated ping. Some solutions surely exist to get a more accurate value.